Movidos modelos relacionados a usuarios de supplies a la app users
